[Feat] 템플릿 단일 선택 로직 적용
