+++ app/src/main/java/com/hara/kaera/presentation/write/TemplateChoiceBottomSheet.kt
@@ -5,6 +5,8 @@ import android.view.View
 import com.hara.kaera.R
 import com.hara.kaera.databinding.BottomsheetTemplateChoiceBinding
 import com.hara.kaera.presentation.base.BindingDraggableBottomSheet
+import com.hara.kaera.presentation.write.adapter.TemplateChoiceAdapter
+import com.hara.kaera.presentation.write.data.DummyTemplateData
 
 class TemplateChoiceBottomSheet :
     BindingDraggableBottomSheet<BottomsheetTemplateChoiceBinding>(R.layout.bottomsheet_template_choice) {
