+++ app/src/main/java/com/hara/kaera/presentation/write/data/TemplateData.kt
@@ -1,4 +1,4 @@
-package com.hara.kaera.presentation.write
+package com.hara.kaera.presentation.write.data
 
 /*
 서버 명세서 기반!
@@ -7,5 +7,5 @@ data class TemplateData(
     val templateId: Int,
     val title: String,
     val info: String,
-    var hasUsed: Boolean
+    val hasUsed: Boolean,
 )
