+++ app/src/main/res/drawable/shape_rect_gray1_stroke_yellow1_8.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/gray_1" />
+    <stroke
+        android:width="1dp"
+        android:color="@color/yellow1" />
+    <corners android:radius="8dp" />
+</shape>
